% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlisence.R
\name{get_license_info}
\alias{get_license_info}
\title{Obtain license information from installed R packages.}
\usage{
get_license_info()
}
\value{
license_df Data.Frame with Package name and license.
}
\description{
Obtain license information from installed R packages.
}

